+++ ofdft_ml/quantum/utils.py
@@ -20,10 +20,10 @@ def xpotential_gen(n_points, n_Gauss, low_a, high_a, low_b, high_b, low_c, high_
         yield(H, Vx)
 
 # periodic potential in k space
-def kpotential_gen(nbasis, n_cos, low_V0, high_V0, low_Phi0, high_Phi0, random_state):
+def kpotential_gen(nbasis, n_gauss, low_V0, high_V0, low_mu, high_mu, low_l, high_l, random_state):
     """
-    Generate periodic potential as sum of cosin functions, differ in their
-    magnitude and phase.
+    Generate periodic potential as sum of gauss-like functions, differ in their
+    magnitude, mean & length scale.
 
     :param nbasis: number of plane wave basis, related to the size of hamilton matrix
     :param n_cos: number of cosin function used
@@ -38,38 +38,23 @@ def kpotential_gen(nbasis, n_cos, low_V0, high_V0, low_Phi0, high_Phi0, random_s
     >>> from quantum.utils import simple_potential_gen
     >>> n_basis = 10
     >>> low_V0, high_V0 = 3, 5
-    >>> low_phi0, high_phi0 = -0.1, 0.1
-    >>> params_gen = simple_potential_gen(n_basis, low_V0, high_V0, low_phi0, high_phi0, random_state=9302)
+    >>> low_mu, high_mu = -0.1, 0.1
+    >>> low_l, high_l = 0.03, 0.1
+    >>> params_gen = kpotential_gen(n_basis, low_V0, high_V0, low_phi0, high_phi0, random_state=9302)
     >>> hamilton_mat, Vq = next(params_gen)
     """
     R = np.random.RandomState(random_state)
     while True:
+        V0 = R.uniform(low_V0, high_V0, n_gauss)
+        Mu = R.uniform(low_mu, high_mu, n_gauss)
+        L = R.uniform(low_l, high_l, n_gauss)
         Vq = np.zeros(nbasis, dtype=np.complex64)
-        hamilton_mat = np.zeros((nbasis, nbasis), dtype=np.complex64)
-
-        V0 = R.uniform(low_V0, high_V0, n_cos)
-        Phi0 = R.uniform(low_Phi0, high_Phi0, n_cos)
-
-        Vq[0] = -np.sum(V0)
-        Vq[1] = 0.5*(V0 @ np.exp(2j*np.pi*Phi0))
-        np.fill_diagonal(hamilton_mat[1:, :-1], Vq[1].conj())
-
-        yield(hamilton_mat, Vq)
-
-# potential in k with two peaks
-def special_potential_gen(nbasis, low_a, high_a, b1_range, b2_range, low_c, high_c, random_state):
-    R = np.random.RandomState(random_state)
-    X = np.linspace(0, 1, 100, endpoint=False)
-    while True:
-        a = R.uniform(low_a, high_a, 2)
-        b = np.append(R.uniform(*b1_range, 1), R.uniform(*b2_range, 1)).reshape((2, 1))
-        c = R.uniform(low_c, high_c, (2, 1))
-        Vx = -a @ np.exp(-(X - b)**2/(2*c**2))
-        Vq = np.fft.rfft(Vx)/100
-        Vq_ = Vq[:nbasis]
+        k_grids = np.arange(0, nbasis, 1)
+        for i in range(n_gauss):
+            Vq += -np.sqrt(2*np.pi)*V0[i]*L[i]*np.exp(-2*(np.pi*L[i]*k_grids)**2)*np.exp(-2j*np.pi*Mu[i]*k_grids)
 
         hamilton_mat = np.zeros((nbasis, nbasis), dtype=np.complex64)
         for i in range(nbasis):
-            np.fill_diagonal(hamilton_mat[i:, :-i], Vq_[i].conj())
-        
-        yield(hamilton_mat, Vq_)
+            np.fill_diagonal(hamilton_mat[i:, :-i], Vq[i].conj())
+
+        yield(hamilton_mat, Vq)